src/requirements.txt
Django==1.8
django-crispy-forms==1.4.0
django-registration-redux==1.1
django-tastypie==0.13.1

